refactor: split config migration modules
